feat: moved toolbar to separate component
