+++ src/app/[category]/components/Toolbar/components/SortBy/index.ts
@@ -1,0 +1,2 @@
+export type { Option } from './SortBy';
+export { default } from './SortBy';
