+++ src/app/[category]/components/Toolbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Toolbar';
