+++ src/app/[category]/page.tsx
@@ -1,12 +1,10 @@
 import { getAllCategories, getCategoryByPath } from '@/shared/api/categoryApi';
 
 import { Breadcrumbs } from '@mantine/core';
-import { Plus } from 'lucide-react';
 import { MDXRemote } from 'next-mdx-remote/rsc';
 import Link from 'next/link';
 import Overview from './components/Overview';
-import SortBy, { type Option } from './components/SortBy/SortBy';
-import Filter from './components/Filter';
+import Toolbar from './components/Toolbar';
 
 export function generateStaticParams() {
   const categories = getAllCategories();
@@ -18,20 +16,14 @@ export function generateStaticParams() {
 
 export const dynamicParams = false;
 
-const sortOptions: Option[] = [
-  { title: 'Featured', value: 'featured' },
-  { title: 'Price, Low to High', value: 'price-asc' },
-  { title: 'Price, High to Low', value: 'price-des' },
-  { title: 'Alphabetically, A - Z', value: 'name-asc' },
-  { title: 'Alphabetically, Z - A', value: 'name-des' },
-];
-
 export default function CatalogPage({
   params,
 }: {
   params: { category: string };
 }) {
   const category = getCategoryByPath(params.category);
+
+  if (!category) return;
 
   return (
     <div className='pb-6 pt-2 md:pb-9 md:pt-4 lg:pb-12'>
@@ -53,12 +45,7 @@ export default function CatalogPage({
           {category?.test_description}
         </p>
 
-        {/* TODO: Move toolbar to separate component */}
-        <div className='mb-4 flex items-center text-sm/4 md:py-2 md:text-base'>
-          <Filter />
-          <p className='hidden md:block'>{category?.productCount} Results</p>
-          <SortBy options={sortOptions} />
-        </div>
+        <Toolbar category={category} />
 
         <p className='mb-[6.1875rem] mt-8 text-center font-light text-brand-grey-700 md:mb-36 md:text-2xl/normal'>
           There are no products in this category yet
